Reorganise step types among files
I thought about continuing "(type)_step.go", but I think that grouping them under a common prefix makes them a little easier to see in file listings.

I also separated input, wait, and scalar (which includes both input and wait), but I'm OK to be pushed back on that.
